chore: rename ui component files
