+++ client/src/app/_components/groups/GroupPostItem.tsx
@@ -12,7 +12,7 @@ import { type UserState } from '~/stores/userStore';
 import { type FirebasePostModel } from '~/models/firebase-post-model';
 import { type IPFSPostModel } from '~/models/ipfs-post-model';
 import { api } from '~/trpc/react';
-import { BasicButton } from '../ui/basic-button';
+import { BasicButton } from '../ui/BasicButton';
 import { FaHeart, FaRegCommentDots } from 'react-icons/fa6';
 import { CiHeart } from 'react-icons/ci';
 import { useWallet } from '~/providers/walletprovider';
@@ -20,7 +20,7 @@ import PostComment from './PostComment';
 import PostCommentList from './PostCommentList';
 import { preventActionNotLoggedIn, preventActionWalletNotConnected } from '~/helpers/user-helper';
 import { toast } from 'react-toastify';
-import ZoomableImage from '../ui/zoomable-image';
+import ZoomableImage from '../ui/ZoomableImage';
 
 const GroupPostItem = (currentPost: FirebasePostModel) => {
 	const { data: postData } = api.PinataPost.getMessage.useQuery({ hash: currentPost.hash });
