+++ client/src/app/_components/ui/ZoomableImage.tsx
