+++ client/src/pages/categories/[categoryId]/index.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router';
-import { PageHeader } from '~/app/_components/ui/page-header';
+import { PageHeader } from '~/app/_components/ui/PageHeader';
 import PlatformLayout from '~/layouts/platform';
 
 export default function Category() {
