+++ client/src/pages/groups/index.tsx
@@ -1,4 +1,4 @@
-import { InlineLink } from '~/app/_components/ui/inline-link';
+import { InlineLink } from '~/app/_components/ui/InlineLink';
 import PlatformLayout from '~/layouts/platform';
 
 export default function Groups() {
