+++ client/src/pages/login/index.tsx
@@ -9,10 +9,10 @@ import { type UserState } from '~/stores/userStore';
 
 import { toast } from 'react-toastify';
 
-import { BasicButton } from '~/app/_components/ui/basic-button';
-import { Checkbox } from '~/app/_components/ui/checkbox';
-import { InlineLink } from '~/app/_components/ui/inline-link';
-import { TextInput } from '~/app/_components/ui/text-input';
+import { BasicButton } from '~/app/_components/ui/BasicButton';
+import { Checkbox } from '~/app/_components/ui/Checkbox';
+import { InlineLink } from '~/app/_components/ui/InlineLink';
+import { TextInput } from '~/app/_components/ui/TextInput';
 
 import { preventActionWalletNotConnected } from '~/helpers/user-helper';
 import DefaultLayout from '~/layouts/default';
