+++ client/src/pages/register/index.tsx
@@ -5,11 +5,11 @@ import { useUserStore } from '~/providers/store-providers/userStoreProvider';
 
 import { toast } from 'react-toastify';
 
-import { BasicButton } from '~/app/_components/ui/basic-button';
-import { Checkbox } from '~/app/_components/ui/checkbox';
-import { InlineLink } from '~/app/_components/ui/inline-link';
-import { SelectOption } from '~/app/_components/ui/select-option';
-import { TextInput } from '~/app/_components/ui/text-input';
+import { BasicButton } from '~/app/_components/ui/BasicButton';
+import { Checkbox } from '~/app/_components/ui/Checkbox';
+import { InlineLink } from '~/app/_components/ui/InlineLink';
+import { SelectOption } from '~/app/_components/ui/SelectOption';
+import { TextInput } from '~/app/_components/ui/TextInput';
 import DefaultLayout from '~/layouts/default';
 
 export default function Register() {
